Add clever player
The clever player looks ahead to see the eventual result of their move, on the assumption that all other players are also clever. The result of pitting two clever players against each other is always a draw.

TODO: we currently assume that it is always best to answer "yes" when asked for a card, as this gives the least information to the opponents. However, we should validate this assumption.